refactor: modularize getRandomElement from loop mode (#33)
