+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - refactor: add proper types for helpers and modes (#33) [`23ac8bc`](https://github.com/henriquehbr/svelte-typewriter/commit/23ac8bc7800166b7cdd4912ad3a092cde6b3ae46)
 - refactor: convert "utils" into separated "helpers" (#33) [`fb6158d`](https://github.com/henriquehbr/svelte-typewriter/commit/fb6158d296e2f99dde8ec7e5a39a6edf39087911)
 - refactor: unify loop and loopRandom modes [`914ef72`](https://github.com/henriquehbr/svelte-typewriter/commit/914ef722e84152f581546cd7e7bba8f168ad144e)
+- refactor: modularize getRandomElement from loop mode (#33) [`8ba2e5f`](https://github.com/henriquehbr/svelte-typewriter/commit/8ba2e5fa77aff40d34884bf0930a6c2f2133ff1c)
 - refactor(utils): integrate loop logic into "writeEffect" (#33) [`1ba9750`](https://github.com/henriquehbr/svelte-typewriter/commit/1ba975028c1e1b068c231fb03967796f7ad0821e)
 - chore: remove unnecessary "any" types [`0e205bd`](https://github.com/henriquehbr/svelte-typewriter/commit/0e205bd2e1fc4b1cca60dfd6f7ada0a1b1cde490)
 - refactor(helpers): add writeEffect [`939a86b`](https://github.com/henriquehbr/svelte-typewriter/commit/939a86b73780d4b2857fdabb1a7f0f2ed49a4416)
@@ -25,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore: bump dependencies [`1444f38`](https://github.com/henriquehbr/svelte-typewriter/commit/1444f38c26cec088998f93d34ec04841408d8b2c)
 - refactor(default): modularize "getLongestTextElement" [`41e7b02`](https://github.com/henriquehbr/svelte-typewriter/commit/41e7b02ffd33914057790a2f04ecd7b2550f5e20)
 - refactor(getLongestTextElement): return Element rather TypewriterElement [`a915357`](https://github.com/henriquehbr/svelte-typewriter/commit/a915357c84cf7da7c61b52e550239c9f267165e9)
-- chore(modes): switch from default to named exports [`1279d4d`](https://github.com/henriquehbr/svelte-typewriter/commit/1279d4de84edac734b1408c2acd9501b454a6406)
+- chore(modes): switch from default to named exports [`119119e`](https://github.com/henriquehbr/svelte-typewriter/commit/119119ea2793c1ce9d835e6a5af3fcb97765db7e)
 - chore: remove duplicated "isInRange" helper types [`951052d`](https://github.com/henriquehbr/svelte-typewriter/commit/951052dbb492c2119b6b987cd43cf14b61b0c4af)
 - chore(cascade): remove unnecessary cursor hiding function [`d26afa6`](https://github.com/henriquehbr/svelte-typewriter/commit/d26afa60818872fce5c704c0aac6f6a5c8ae541e)
