--- lib/component/esm/loop.mjs
@@ -1,1 +1,0 @@
-import"svelte/internal";import{t}from"./Typewriter.mjs";import{o as e,n as o}from"./unwriteEffect.mjs";import{o as r}from"./typingInterval.mjs";import{o as a}from"./writeEffect.mjs";let n=[];const s=t=>{for(;;){const e=r(0,t.length),o="number"==typeof n&&e!==n;if(Array.isArray(n)&&!n.includes(e)||o)return o&&(n=[]),n.push(e),t[e];n.length===t.length&&(n=n.pop())}},i=async(r,{currentNode:n,text:s},i)=>{e(r);const m=n.tagName.toLowerCase(),p=t(s,m);[...n.attributes].forEach((({name:t,value:e})=>p.setAttribute(t,e))),r.appendChild(p),await a({currentNode:p,text:s},i);const f=s.replaceAll("&","&amp;");p.innerHTML===f&&await o(p,i),e(r)},m=async({node:t,elements:e},o)=>{for(;;)if(o.loop)for(const r of e)await i(t,r,o);else if(o.loopRandom){const r=s(e);await i(t,r,o)}};export{m as mode};
