+++ lib/component/esm/loopTypewriter.mjs
@@ -1,0 +1,1 @@
+import"svelte/internal";import{t}from"./Typewriter.mjs";import{t as e,o,n as a}from"./getRandomElement.mjs";import"./typingInterval.mjs";import{o as r}from"./writeEffect.mjs";const n=async(e,{currentNode:n,text:s},m)=>{o(e);const i=n.tagName.toLowerCase(),p=t(s,i);[...n.attributes].forEach((({name:t,value:e})=>p.setAttribute(t,e))),e.appendChild(p),await r({currentNode:p,text:s},m);const l=s.replaceAll("&","&amp;");p.innerHTML===l&&await a(p,m),o(e)},s=async({node:t,elements:o},a)=>{for(;;)if(a.loop)for(const e of o)await n(t,e,a);else if(a.loopRandom){const r=e(o);await n(t,r,a)}};export{s as mode};
