--- lib/component/esm/unwriteEffect.mjs
@@ -1,1 +1,0 @@
-import{s as e}from"./typingInterval.mjs";const o=e=>e.childNodes.forEach((e=>e.remove())),n=async(o,n)=>{n.dispatch("done"),await e("number"==typeof n.loop?n.loop:1500);const t=o.innerHTML.replaceAll("&amp;","&");for(let a=t.length-1;a>=0;a--)">"===t[a]&&(a=t.lastIndexOf("<",a)),o.innerHTML=t.slice(0,a),await e(n.interval)};export{n,o};
