+++ lib/component/umd/index.js
@@ -1,1 +1,1 @@
-!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?module.exports=e():"function"==typeof define&&define.amd?define(e):(t="undefined"!=typeof globalThis?globalThis:t||self).Typewriter=e()}(this,(function(){"use strict";function t(){}function e(t){return t()}function n(){return Object.create(null)}function o(t){t.forEach(e)}function r(t){return"function"==typeof t}function c(t,e){return t!=t?e==e:t!==e||t&&"object"==typeof t||"function"==typeof t}function s(t,e,n,o){return t[1]&&o?function(t,e){for(const n in e)t[n]=e[n];return t}(n.ctx.slice(),t[1](o(e))):n.ctx}function i(t,e,n,o,r,c,i){const a=function(t,e,n,o){if(t[2]&&o){const r=t[2](o(n));if(void 0===e.dirty)return r;if("object"==typeof r){const t=[],n=Math.max(e.dirty.length,r.length);for(let o=0;o<n;o+=1)t[o]=e.dirty[o]|r[o];return t}return e.dirty|r}return e.dirty}(e,o,r,c);if(a){const r=s(e,n,o,i);t.p(r,a)}}function a(t){t.parentNode.removeChild(t)}function l(t){return document.createElement(t)}function u(t,e,n,o){t.style.setProperty(e,n,o?"important":"")}function d(t,e,n){t.classList[n?"add":"remove"](e)}let f;function p(t){f=t}function m(){const t=function(){if(!f)throw new Error("Function called outside component initialization");return f}();return(e,n)=>{const o=t.$$.callbacks[e];if(o){const r=function(t,e){const n=document.createEvent("CustomEvent");return n.initCustomEvent(t,!1,!1,e),n}(e,n);o.slice().forEach((e=>{e.call(t,r)}))}}}const h=[],y=[],g=[],b=[],v=Promise.resolve();let $=!1;function x(t){g.push(t)}let _=!1;const w=new Set;function E(){if(!_){_=!0;do{for(let t=0;t<h.length;t+=1){const e=h[t];p(e),L(e.$$)}for(p(null),h.length=0;y.length;)y.pop()();for(let t=0;t<g.length;t+=1){const e=g[t];w.has(e)||(w.add(e),e())}g.length=0}while(h.length);for(;b.length;)b.pop()();$=!1,_=!1,w.clear()}}function L(t){if(null!==t.fragment){t.update(),o(t.before_update);const e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(x)}}const T=new Set;function N(t,e){t&&t.i&&(T.delete(t),t.i(e))}function M(t,e){-1===t.$$.dirty[0]&&(h.push(t),$||($=!0,v.then(E)),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function C(c,s,i,l,u,d,m=[-1]){const h=f;p(c);const y=s.props||{},g=c.$$={fragment:null,ctx:null,props:d,update:t,not_equal:u,bound:n(),on_mount:[],on_destroy:[],before_update:[],after_update:[],context:new Map(h?h.$$.context:[]),callbacks:n(),dirty:m,skip_bound:!1};let b=!1;if(g.ctx=i?i(c,y,((t,e,...n)=>{const o=n.length?n[0]:e;return g.ctx&&u(g.ctx[t],g.ctx[t]=o)&&(!g.skip_bound&&g.bound[t]&&g.bound[t](o),b&&M(c,t)),e})):[],g.update(),b=!0,o(g.before_update),g.fragment=!!l&&l(g.ctx),s.target){if(s.hydrate){const t=function(t){return Array.from(t.childNodes)}(s.target);g.fragment&&g.fragment.l(t),t.forEach(a)}else g.fragment&&g.fragment.c();s.intro&&N(c.$$.fragment),function(t,n,c){const{fragment:s,on_mount:i,on_destroy:a,after_update:l}=t.$$;s&&s.m(n,c),x((()=>{const n=i.map(e).filter(r);a?a.push(...n):o(n),t.$$.on_mount=[]})),l.forEach(x)}(c,s.target,s.anchor),E()}p(h)}const A=(t,e)=>{const n=document.createElement(e);return n.textContent=t,n},k=t=>1===t.childNodes.length&&3===t.childNodes[0].nodeType,H=t=>{if(k(t)){const e=t.textContent,n=A(t.textContent,"p");return t.textContent="",t.appendChild(n),[{currentNode:n,text:e}]}return[...t.children].map((t=>{const e=t.innerHTML.replaceAll("&amp;","&");return{currentNode:t,text:e}}))},O=async(t,e)=>{const{mode:n}=(e.loop||e.loopRandom)&&await Promise.resolve().then((function(){return K}))||e.scramble&&await Promise.resolve().then((function(){return Y}))||await Promise.resolve().then((function(){return tt})),o=H(t);if(e.delay>0){const{sleep:n}=await Promise.resolve().then((function(){return nt}));await n(e.delay),t.classList.remove("delay")}n({node:t,elements:o},e)};function j(e){let n,o,c,f,p;const m=e[9].default,h=function(t,e,n,o){if(t){const r=s(t,e,n,o);return t[0](r)}}(m,e,e[8],null);return{c(){var t,o,r;n=l("div"),h&&h.c(),u(n,"--cursor-color","string"==typeof e[0]?e[0]:"black"),t=n,o="class",null==(r="svelte-8gtb9p")?t.removeAttribute(o):t.getAttribute(o)!==r&&t.setAttribute(o,r),d(n,"cursor",e[0]),d(n,"delay",e[1].delay>0)},m(s,i){var a;!function(t,e,n){t.insertBefore(e,n||null)}(s,n,i),h&&h.m(n,null),c=!0,f||(a=o=O.call(null,n,e[1]),p=a&&r(a.destroy)?a.destroy:t,f=!0)},p(t,[e]){h&&h.p&&256&e&&i(h,m,t,t[8],e,null,null),(!c||1&e)&&u(n,"--cursor-color","string"==typeof t[0]?t[0]:"black"),1&e&&d(n,"cursor",t[0]),2&e&&d(n,"delay",t[1].delay>0)},i(t){c||(N(h,t),c=!0)},o(t){!function(t,e,n,o){if(t&&t.o){if(T.has(t))return;T.add(t),(void 0).c.push((()=>{T.delete(t),o&&(n&&t.d(1),o())})),t.o(e)}}(h,t),c=!1},d(t){t&&a(n),h&&h.d(t),f=!1,p()}}}function P(t,e,n){let{$$slots:o={},$$scope:r}=e,{interval:c=30}=e,{cascade:s=!1}=e,{loop:i=!1}=e,{loopRandom:a=!1}=e,{scramble:l=!1}=e,{cursor:u=!0}=e,{delay:d=0}=e;const f=m(),p={interval:c,cascade:s,loop:i,loopRandom:a,scramble:l,cursor:u,delay:d,dispatch:f};return t.$$set=t=>{"interval"in t&&n(2,c=t.interval),"cascade"in t&&n(3,s=t.cascade),"loop"in t&&n(4,i=t.loop),"loopRandom"in t&&n(5,a=t.loopRandom),"scramble"in t&&n(6,l=t.scramble),"cursor"in t&&n(0,u=t.cursor),"delay"in t&&n(7,d=t.delay),"$$scope"in t&&n(8,r=t.$$scope)},[u,p,c,s,i,a,l,d,r,o]}const R=t=>t.childNodes.forEach((t=>t.remove())),z=(t,e)=>Math.floor(Math.random()*(e-t)+t),I=t=>new Promise((e=>setTimeout(e,t))),S=async t=>I(Array.isArray(t)?t[z(0,t.length)]:t),F=async(t,e)=>{e.dispatch("done"),await S("number"==typeof e.loop?e.loop:1500);const n=t.innerHTML.replaceAll("&amp;","&");for(let o=n.length-1;o>=0;o--)">"===n[o]&&(o=n.lastIndexOf("<",o)),t.innerHTML=n.slice(0,o),await S(e.interval)},B=t=>null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing"),D=async({currentNode:t,text:e},n)=>{t.classList.add("typing");for(let o=0;o<=e.length;o++)"<"===e[o]&&(o=e.indexOf(">",o)),t.innerHTML=e.slice(0,o),await S(n.interval);!n.loop&&!n.loopRandom&&B(t)};let q=[];const G=t=>{for(;;){const e=z(0,t.length),n="number"==typeof q&&e!==q;if(Array.isArray(q)&&!q.includes(e)||n)return n&&(q=[]),q.push(e),t[e];q.length===t.length&&(q=q.pop())}},J=async(t,{currentNode:e,text:n},o)=>{R(t);const r=e.tagName.toLowerCase(),c=A(n,r);[...e.attributes].forEach((({name:t,value:e})=>c.setAttribute(t,e))),t.appendChild(c),await D({currentNode:c,text:n},o);const s=n.replaceAll("&","&amp;");c.innerHTML===s&&await F(c,o),R(t)};var K=Object.freeze({__proto__:null,mode:async({node:t,elements:e},n)=>{for(;;)if(n.loop)for(const o of e)await J(t,o,n);else if(n.loopRandom){const o=G(e);await J(t,o,n)}}});const Q=(t,[e,n])=>t>=e&&t<=n;let U=[];const V=t=>U.find((e=>e.currentNode===t)),W=t=>{const e=(t=>{const e=/(<([^>]+)>)/g,n=[];let o;for(;null!==(o=e.exec(t.innerHTML));){const t=o.index,r=e.lastIndex;n.push([t,r-1])}return n})(t),n=t.innerHTML.split("").map(((n,o)=>{const{matchingLetters:r}=V(t),c=r.includes(o)||n.match(/\s+/g)||((t,e)=>e.some((([e,n])=>Q(t,[e,n]))))(o,e),s=String.fromCharCode(65+Math.round(50*Math.random()));return c?n:s})).join("");t.innerHTML=n},X=(t,e)=>{const n=t.innerHTML;for(let o=0;o<e.length;o++){const r=e[o],{matchingLetters:c}=V(t);!c.includes(o)&&r===n[o]&&c.push(o)}};var Y=Object.freeze({__proto__:null,mode:async({elements:t},e)=>{U=[...t.map((({currentNode:t})=>({currentNode:t,matchingLetters:[]})))],await new Promise((n=>{t.forEach((async({currentNode:t,text:o})=>{const r="number"==typeof e.scramble?e.scramble:3e3,c=(new Date).getTime();for(;;){W(t),X(t,o),await S(e.interval);const s=t.innerHTML!=o,i=(new Date).getTime()-c<r;if(!s||!i){n();break}}t.innerHTML=o}))})),e.dispatch("done")}});const Z=(t,e)=>{new MutationObserver((t=>{t.forEach((t=>{const n="attributes"===t.type,o=!t.target.classList.contains("typing");n&&o&&e()}))})).observe(t,{attributes:!0,childList:!0,subtree:!0})};var tt=Object.freeze({__proto__:null,mode:async({elements:t},e)=>{if(e.cascade)(t=>{t.forEach((t=>t.currentNode.textContent=""))})(t);else{const{getLongestTextElement:n}=await Promise.resolve().then((function(){return nt})),o=n(t);Z(o,(()=>e.dispatch("done")))}for(const n of t)e.cascade?await D(n,e):D(n,e);e.cascade&&e.dispatch("done")}});const et=(t,e)=>e.text.length-t.text.length;var nt=Object.freeze({__proto__:null,cleanChildNodes:R,createElement:A,getElements:H,hasSingleTextNode:k,rng:z,sleep:I,typingInterval:S,unwriteEffect:F,isInRange:Q,writeEffect:D,hideCursorOnAnimationEnd:B,getLongestTextElement:t=>t.sort(et)[0].currentNode,onAnimationEnd:Z});return class extends class{$destroy(){!function(t,e){const n=t.$$;null!==n.fragment&&(o(n.on_destroy),n.fragment&&n.fragment.d(e),n.on_destroy=n.fragment=null,n.ctx=[])}(this,1),this.$destroy=t}$on(t,e){const n=this.$$.callbacks[t]||(this.$$.callbacks[t]=[]);return n.push(e),()=>{const t=n.indexOf(e);-1!==t&&n.splice(t,1)}}$set(t){var e;this.$$set&&(e=t,0!==Object.keys(e).length)&&(this.$$.skip_bound=!0,this.$$set(t),this.$$.skip_bound=!1)}}{constructor(t){var e,n,o;super(),document.getElementById("svelte-8gtb9p-style")||((o=l("style")).id="svelte-8gtb9p-style",o.textContent="@keyframes svelte-8gtb9p-cursorFade{0%,100%{opacity:1}50%{opacity:0}}.cursor.svelte-8gtb9p .typing::after{content:'▌';display:inline-block;color:var(--cursor-color);animation:svelte-8gtb9p-cursorFade 1.25s infinite}.delay.svelte-8gtb9p{visibility:hidden}",e=document.head,n=o,e.appendChild(n)),C(this,t,P,j,c,{interval:2,cascade:3,loop:4,loopRandom:5,scramble:6,cursor:0,delay:7})}}}));
+!function(t,e){"object"==typeof exports&&"undefined"!=typeof module?module.exports=e():"function"==typeof define&&define.amd?define(e):(t="undefined"!=typeof globalThis?globalThis:t||self).Typewriter=e()}(this,(function(){"use strict";function t(){}function e(t){return t()}function n(){return Object.create(null)}function o(t){t.forEach(e)}function r(t){return"function"==typeof t}function c(t,e){return t!=t?e==e:t!==e||t&&"object"==typeof t||"function"==typeof t}function s(t,e,n,o){return t[1]&&o?function(t,e){for(const n in e)t[n]=e[n];return t}(n.ctx.slice(),t[1](o(e))):n.ctx}function a(t,e,n,o,r,c,a){const i=function(t,e,n,o){if(t[2]&&o){const r=t[2](o(n));if(void 0===e.dirty)return r;if("object"==typeof r){const t=[],n=Math.max(e.dirty.length,r.length);for(let o=0;o<n;o+=1)t[o]=e.dirty[o]|r[o];return t}return e.dirty|r}return e.dirty}(e,o,r,c);if(i){const r=s(e,n,o,a);t.p(r,i)}}function i(t){t.parentNode.removeChild(t)}function l(t){return document.createElement(t)}function u(t,e,n,o){t.style.setProperty(e,n,o?"important":"")}function d(t,e,n){t.classList[n?"add":"remove"](e)}let f;function p(t){f=t}function m(){const t=function(){if(!f)throw new Error("Function called outside component initialization");return f}();return(e,n)=>{const o=t.$$.callbacks[e];if(o){const r=function(t,e){const n=document.createEvent("CustomEvent");return n.initCustomEvent(t,!1,!1,e),n}(e,n);o.slice().forEach((e=>{e.call(t,r)}))}}}const h=[],y=[],g=[],b=[],v=Promise.resolve();let $=!1;function x(t){g.push(t)}let _=!1;const w=new Set;function E(){if(!_){_=!0;do{for(let t=0;t<h.length;t+=1){const e=h[t];p(e),L(e.$$)}for(p(null),h.length=0;y.length;)y.pop()();for(let t=0;t<g.length;t+=1){const e=g[t];w.has(e)||(w.add(e),e())}g.length=0}while(h.length);for(;b.length;)b.pop()();$=!1,_=!1,w.clear()}}function L(t){if(null!==t.fragment){t.update(),o(t.before_update);const e=t.dirty;t.dirty=[-1],t.fragment&&t.fragment.p(t.ctx,e),t.after_update.forEach(x)}}const T=new Set;function N(t,e){t&&t.i&&(T.delete(t),t.i(e))}function M(t,e){-1===t.$$.dirty[0]&&(h.push(t),$||($=!0,v.then(E)),t.$$.dirty.fill(0)),t.$$.dirty[e/31|0]|=1<<e%31}function C(c,s,a,l,u,d,m=[-1]){const h=f;p(c);const y=s.props||{},g=c.$$={fragment:null,ctx:null,props:d,update:t,not_equal:u,bound:n(),on_mount:[],on_destroy:[],before_update:[],after_update:[],context:new Map(h?h.$$.context:[]),callbacks:n(),dirty:m,skip_bound:!1};let b=!1;if(g.ctx=a?a(c,y,((t,e,...n)=>{const o=n.length?n[0]:e;return g.ctx&&u(g.ctx[t],g.ctx[t]=o)&&(!g.skip_bound&&g.bound[t]&&g.bound[t](o),b&&M(c,t)),e})):[],g.update(),b=!0,o(g.before_update),g.fragment=!!l&&l(g.ctx),s.target){if(s.hydrate){const t=function(t){return Array.from(t.childNodes)}(s.target);g.fragment&&g.fragment.l(t),t.forEach(i)}else g.fragment&&g.fragment.c();s.intro&&N(c.$$.fragment),function(t,n,c){const{fragment:s,on_mount:a,on_destroy:i,after_update:l}=t.$$;s&&s.m(n,c),x((()=>{const n=a.map(e).filter(r);i?i.push(...n):o(n),t.$$.on_mount=[]})),l.forEach(x)}(c,s.target,s.anchor),E()}p(h)}const A=(t,e)=>{const n=document.createElement(e);return n.textContent=t,n},k=t=>1===t.childNodes.length&&3===t.childNodes[0].nodeType,H=t=>{if(k(t)){const e=t.textContent,n=A(t.textContent,"p");return t.textContent="",t.appendChild(n),[{currentNode:n,text:e}]}return[...t.children].map((t=>{const e=t.innerHTML.replaceAll("&amp;","&");return{currentNode:t,text:e}}))},O=async(t,e)=>{const{mode:n}=(e.loop||e.loopRandom)&&await Promise.resolve().then((function(){return J}))||e.scramble&&await Promise.resolve().then((function(){return X}))||await Promise.resolve().then((function(){return Z})),o=H(t);if(e.delay>0){const{sleep:n}=await Promise.resolve().then((function(){return et}));await n(e.delay),t.classList.remove("delay")}n({node:t,elements:o},e)};function j(e){let n,o,c,f,p;const m=e[9].default,h=function(t,e,n,o){if(t){const r=s(t,e,n,o);return t[0](r)}}(m,e,e[8],null);return{c(){var t,o,r;n=l("div"),h&&h.c(),u(n,"--cursor-color","string"==typeof e[0]?e[0]:"black"),t=n,o="class",null==(r="svelte-8gtb9p")?t.removeAttribute(o):t.getAttribute(o)!==r&&t.setAttribute(o,r),d(n,"cursor",e[0]),d(n,"delay",e[1].delay>0)},m(s,a){var i;!function(t,e,n){t.insertBefore(e,n||null)}(s,n,a),h&&h.m(n,null),c=!0,f||(i=o=O.call(null,n,e[1]),p=i&&r(i.destroy)?i.destroy:t,f=!0)},p(t,[e]){h&&h.p&&256&e&&a(h,m,t,t[8],e,null,null),(!c||1&e)&&u(n,"--cursor-color","string"==typeof t[0]?t[0]:"black"),1&e&&d(n,"cursor",t[0]),2&e&&d(n,"delay",t[1].delay>0)},i(t){c||(N(h,t),c=!0)},o(t){!function(t,e,n,o){if(t&&t.o){if(T.has(t))return;T.add(t),(void 0).c.push((()=>{T.delete(t),o&&(n&&t.d(1),o())})),t.o(e)}}(h,t),c=!1},d(t){t&&i(n),h&&h.d(t),f=!1,p()}}}function R(t,e,n){let{$$slots:o={},$$scope:r}=e,{interval:c=30}=e,{cascade:s=!1}=e,{loop:a=!1}=e,{loopRandom:i=!1}=e,{scramble:l=!1}=e,{cursor:u=!0}=e,{delay:d=0}=e;const f=m(),p={interval:c,cascade:s,loop:a,loopRandom:i,scramble:l,cursor:u,delay:d,dispatch:f};return t.$$set=t=>{"interval"in t&&n(2,c=t.interval),"cascade"in t&&n(3,s=t.cascade),"loop"in t&&n(4,a=t.loop),"loopRandom"in t&&n(5,i=t.loopRandom),"scramble"in t&&n(6,l=t.scramble),"cursor"in t&&n(0,u=t.cursor),"delay"in t&&n(7,d=t.delay),"$$scope"in t&&n(8,r=t.$$scope)},[u,p,c,s,a,i,l,d,r,o]}const P=t=>t.childNodes.forEach((t=>t.remove())),z=(t,e)=>Math.floor(Math.random()*(e-t)+t),I=t=>new Promise((e=>setTimeout(e,t))),S=async t=>I(Array.isArray(t)?t[z(0,t.length)]:t),F=async(t,e)=>{e.dispatch("done"),await S("number"==typeof e.loop?e.loop:1500);const n=t.innerHTML.replaceAll("&amp;","&");for(let o=n.length-1;o>=0;o--)">"===n[o]&&(o=n.lastIndexOf("<",o)),t.innerHTML=n.slice(0,o),await S(e.interval)},B=t=>null!==t.nextSibling&&1==t.classList.length?t.removeAttribute("class"):t.classList.remove("typing"),D=async({currentNode:t,text:e},n)=>{t.classList.add("typing");for(let o=0;o<=e.length;o++)"<"===e[o]&&(o=e.indexOf(">",o)),t.innerHTML=e.slice(0,o),await S(n.interval);!n.loop&&!n.loopRandom&&B(t)},q=t=>{let e=[];for(;;){const n=z(0,t.length),o="number"==typeof e&&n!==e;if(Array.isArray(e)&&!e.includes(n)||o)return o&&(e=[]),e.push(n),t[n];e.length===t.length&&(e=e.pop())}},G=async(t,{currentNode:e,text:n},o)=>{P(t);const r=e.tagName.toLowerCase(),c=A(n,r);[...e.attributes].forEach((({name:t,value:e})=>c.setAttribute(t,e))),t.appendChild(c),await D({currentNode:c,text:n},o);const s=n.replaceAll("&","&amp;");c.innerHTML===s&&await F(c,o),P(t)};var J=Object.freeze({__proto__:null,mode:async({node:t,elements:e},n)=>{for(;;)if(n.loop)for(const o of e)await G(t,o,n);else if(n.loopRandom){const o=q(e);await G(t,o,n)}}});const K=(t,[e,n])=>t>=e&&t<=n;let Q=[];const U=t=>Q.find((e=>e.currentNode===t)),V=t=>{const e=(t=>{const e=/(<([^>]+)>)/g,n=[];let o;for(;null!==(o=e.exec(t.innerHTML));){const t=o.index,r=e.lastIndex;n.push([t,r-1])}return n})(t),n=t.innerHTML.split("").map(((n,o)=>{const{matchingLetters:r}=U(t),c=r.includes(o)||n.match(/\s+/g)||((t,e)=>e.some((([e,n])=>K(t,[e,n]))))(o,e),s=String.fromCharCode(65+Math.round(50*Math.random()));return c?n:s})).join("");t.innerHTML=n},W=(t,e)=>{const n=t.innerHTML;for(let o=0;o<e.length;o++){const r=e[o],{matchingLetters:c}=U(t);!c.includes(o)&&r===n[o]&&c.push(o)}};var X=Object.freeze({__proto__:null,mode:async({elements:t},e)=>{Q=[...t.map((({currentNode:t})=>({currentNode:t,matchingLetters:[]})))],await new Promise((n=>{t.forEach((async({currentNode:t,text:o})=>{const r="number"==typeof e.scramble?e.scramble:3e3,c=(new Date).getTime();for(;;){V(t),W(t,o),await S(e.interval);const s=t.innerHTML!=o,a=(new Date).getTime()-c<r;if(!s||!a){n();break}}t.innerHTML=o}))})),e.dispatch("done")}});const Y=(t,e)=>{new MutationObserver((t=>{t.forEach((t=>{const n="attributes"===t.type,o=!t.target.classList.contains("typing");n&&o&&e()}))})).observe(t,{attributes:!0,childList:!0,subtree:!0})};var Z=Object.freeze({__proto__:null,mode:async({elements:t},e)=>{if(e.cascade)(t=>{t.forEach((t=>t.currentNode.textContent=""))})(t);else{const{getLongestTextElement:n}=await Promise.resolve().then((function(){return et})),o=n(t);Y(o,(()=>e.dispatch("done")))}for(const n of t)e.cascade?await D(n,e):D(n,e);e.cascade&&e.dispatch("done")}});const tt=(t,e)=>e.text.length-t.text.length;var et=Object.freeze({__proto__:null,cleanChildNodes:P,createElement:A,getElements:H,hasSingleTextNode:k,rng:z,sleep:I,typingInterval:S,unwriteEffect:F,isInRange:K,writeEffect:D,hideCursorOnAnimationEnd:B,getLongestTextElement:t=>t.sort(tt)[0].currentNode,onAnimationEnd:Y,getRandomElement:q});return class extends class{$destroy(){!function(t,e){const n=t.$$;null!==n.fragment&&(o(n.on_destroy),n.fragment&&n.fragment.d(e),n.on_destroy=n.fragment=null,n.ctx=[])}(this,1),this.$destroy=t}$on(t,e){const n=this.$$.callbacks[t]||(this.$$.callbacks[t]=[]);return n.push(e),()=>{const t=n.indexOf(e);-1!==t&&n.splice(t,1)}}$set(t){var e;this.$$set&&(e=t,0!==Object.keys(e).length)&&(this.$$.skip_bound=!0,this.$$set(t),this.$$.skip_bound=!1)}}{constructor(t){var e,n,o;super(),document.getElementById("svelte-8gtb9p-style")||((o=l("style")).id="svelte-8gtb9p-style",o.textContent="@keyframes svelte-8gtb9p-cursorFade{0%,100%{opacity:1}50%{opacity:0}}.cursor.svelte-8gtb9p .typing::after{content:'▌';display:inline-block;color:var(--cursor-color);animation:svelte-8gtb9p-cursorFade 1.25s infinite}.delay.svelte-8gtb9p{visibility:hidden}",e=document.head,n=o,e.appendChild(n)),C(this,t,R,j,c,{interval:2,cascade:3,loop:4,loopRandom:5,scramble:6,cursor:0,delay:7})}}}));
