+++ lib/core/helpers/getRandomElement.js
@@ -1,0 +1,1 @@
+import{rng as r}from"./rng.js";const t=t=>{let e=[];for(;;){const n=r(0,t.length),o="number"==typeof e&&n!==e;if(Array.isArray(e)&&!e.includes(n)||o){o&&(e=[]),e.push(n);return t[n]}e.length===t.length&&(e=e.pop())}};export{t as getRandomElement};
