+++ lib/core/modes/index.js
@@ -1,1 +1,1 @@
-import{getElements as e}from"../helpers/getElements.js";const o=async(o,t)=>{const{mode:s}=(t.loop||t.loopRandom)&&await import("./loop.js")||t.scramble&&await import("./scramble.js")||await import("./typewriter.js"),a=e(o);if(t.delay>0){const{sleep:e}=await import("../helpers/index.js");await e(t.delay),o.classList.remove("delay")}s({node:o,elements:a},t)};export{o as typewriter};
+import{getElements as e}from"../helpers/getElements.js";const t=async(t,o)=>{const{mode:s}=(o.loop||o.loopRandom)&&await import("./loopTypewriter.js")||o.scramble&&await import("./scramble.js")||await import("./typewriter.js"),a=e(t);if(o.delay>0){const{sleep:e}=await import("../helpers/index.js");await e(o.delay),t.classList.remove("delay")}s({node:t,elements:a},o)};export{t as typewriter};
