+++ lib/core/modes/loopTypewriter.js
@@ -1,0 +1,1 @@
+import{cleanChildNodes as e}from"../helpers/cleanChildNodes.js";import{createElement as t}from"../helpers/createElement.js";import{unwriteEffect as o}from"../helpers/unwriteEffect.js";import{writeEffect as r}from"../helpers/writeEffect.js";import{getRandomElement as a}from"../helpers/getRandomElement.js";const s=async(a,{currentNode:s,text:n},i)=>{e(a);const l=s.tagName.toLowerCase(),m=t(n,l);[...s.attributes].forEach((({name:e,value:t})=>m.setAttribute(e,t))),a.appendChild(m),await r({currentNode:m,text:n},i);const p=n.replaceAll("&","&amp;");m.innerHTML===p&&await o(m,i),e(a)},n=async({node:e,elements:t},o)=>{for(;;)if(o.loop)for(const r of t)await s(e,r,o);else if(o.loopRandom){const r=a(t);await s(e,r,o)}};export{n as mode};
