+++ src/helpers/index.ts
@@ -12,3 +12,4 @@ export * from './writeEffect'
 export * from './hideCursorOnAnimationEnd'
 export * from './getLongestTextElement'
 export * from './onAnimationEnd'
+export * from './getRandomElement'
