+++ src/modes/index.ts
@@ -3,7 +3,7 @@ import type { TypewriterMainFn } from '@svelte-typewriter/types'
 
 const typewriter: TypewriterMainFn = async (node, options) => {
 	const { mode } =
-		((options.loop || options.loopRandom) && (await import('./loop'))) ||
+		((options.loop || options.loopRandom) && (await import('./loopTypewriter'))) ||
 		(options.scramble && (await import('./scramble'))) ||
 		(await import('./typewriter'))
 	const elements = getElements(node)
