+++ src/types/loop.ts
@@ -1,11 +1,9 @@
 import type { TypewriterElement, TypewriterOptions } from './typewriter'
 
-type GetRandomText = (element: TypewriterElement[]) => TypewriterElement
-
 type LoopAnimation = (
 	parentElement: Element,
 	element: TypewriterElement,
 	options: TypewriterOptions
 ) => Promise<void>
 
-export { GetRandomText, LoopAnimation }
+export { LoopAnimation }
